SEGF OK, LEAKS OK, nettoyage ficher, creation srcs/tools pour gestion de path
