+++ includes/sh.h
@@ -16,10 +16,18 @@ typedef struct	s_cfg t_cfg;
 
 void	ex(char *s); //pour la gestion des erreurs amenant a quiter
 
-void	ft_ex(char *error);
-void	set_signal_ign(void);
-uint8_t	init_shell(char **env, char **av);
-t_cfg	*init_cfg(char **env, char **av);
-t_cfg	*cfg_shell(void);
+void		ft_ex(char *error);
+void		clean_cfg(t_cfg *shell);
+void		set_signal_ign(void);
+uint8_t		init_shell(char **env, char **av);
+t_cfg		*init_cfg(char **env, char **av);
+t_cfg		*cfg_shell(void);
+
+/*	COMMON TOOLS */
+uint32_t	path_errors(char *path, uint8_t check_it);
+uint8_t		c_enametoolong(char *path);
+uint8_t		c_isdir(char *path);
+char		*create_abs_path(char *s);
+uint32_t	check_access(char *path, int right);
 
 #endif
