+++ srcs/builtins/unsetenv.c
@@ -5,7 +5,7 @@
 #include "var.h"
 
 int			unsetenv_find(void *elem, void *var)
- {
+{
 	 if (ft_strcmp(((t_var *)elem)->ctab[0], (char *)var) == 0)
         return (1);
     return (0);
@@ -13,7 +13,7 @@ int			unsetenv_find(void *elem, void *var)
 
 void		unsetenv_del(void *delete, size_t size)
 {
-	(void)size; //pk avoir rajouter the fcking size ?
+	(void)size;
     ft_del_tab(((void **)((t_var *)delete)->ctab));
     free(delete);
 }
@@ -22,7 +22,7 @@ uint8_t			ft_unsetenv(t_job *j, t_process *p)
 {
 	int		i;
 
-	(void)j; //useles job pour tout le monde ????????????
+	(void)j;
 	i = 0;
 	while (p->av[++i])
 		ft_lstdelif(&cfg_shell()->env, p->av[i], unsetenv_find, unsetenv_del);
