+++ srcs/evaluator/constructor.c
@@ -23,17 +23,16 @@ void	cmd_to_process(t_list **lst, t_list *s_cmd)
 		ft_lst_push_back(lst, &process, sizeof(process));
 		s_cmd = s_cmd->next;
 	}
-	return ;   /*i useless sans create_pipe_list ???*/
+	return ;
 }
 
-int		cmd_to_job(t_cfg *shell, t_job *job, t_list *s_cmd, char *cmd)
+void	cmd_to_job(t_cfg *shell, t_job *job, t_list *s_cmd, char *cmd)
 {
-	ft_bzero(job, sizeof(t_job));   /*set t_pipe a 0. Besoin d etre a -1?*/
+	ft_bzero(job, sizeof(t_job));
 	cmd_to_process(&job->process, s_cmd);
 	job->cmd = ft_strdup(cmd);
 	job->std[0] = dup(STDIN_FILENO);
 	job->std[1] = dup(STDOUT_FILENO);
 	job->std[2] = dup(STDERR_FILENO);
 	term_create_eval(&shell->term_origin, &job->term_eval);
-	return (0);
 }
