+++ srcs/evaluator/exec_type.c
@@ -4,12 +4,6 @@
 #include "var.h"
 #include "sh.h"
 
-
-/*
- *Command Search and Execution
- *https://pubs.opengroup.org/onlinepubs/9699919799/utilities/V3_chap02.html#tag_18_09_01_01
- */
-
 static int			builtin_search(t_process *p)
 {
 	if (!ft_strcmp(p->cmd, "echo"))
@@ -26,7 +20,6 @@ static int			builtin_search(t_process *p)
 		return (p->setup |= B_EXIT);
 	if (!ft_strcmp(p->cmd, "hash"))
 		return (p->setup |= B_HASH);
-
 	return (0);
 }
 
@@ -46,28 +39,14 @@ static uint16_t		find_type(t_list *env, t_process *p)
 	return (SUCCESS);
 }
 
-
-static void		any_slash(t_list *env, t_process *p)
+static void			any_slash(t_list *env, t_process *p)
 {
 	if (find_type(env, p) || p->setup & BUILTIN)
 		return ;
 	p->setup |= path_errors(p->path, 1);
 }
 
-
-char		*create_abs_path(char *s)
-{
-	char		*buf;
-	char		*tmp;
-
-	if (!(buf = getcwd(0, 0)))
-		perror("getcwd");
-	tmp = ft_strjoin(3, buf, "/", s);
-	ft_strdel(&buf);
-	return (tmp);
-}
-
-static void		with_slash(t_process *p)
+static void			with_slash(t_process *p)
 {
 	char		*tmp;
 
@@ -83,7 +62,7 @@ static void		with_slash(t_process *p)
 	}
 }
 
-void	process_type(t_process *p)
+void				process_type(t_process *p)
 {
 	if (ft_strchr(p->cmd, '/'))
 		with_slash(p);
