+++ srcs/evaluator/leveling.c
@@ -1,12 +1,10 @@
 #include "libft.h"
+#include "ft_printf.h"
 #include "exec.h"
 #include "parser.h"
 #include "sh.h"
 
-
-#include <stdio.h>   //debug
-
-int		condition_respectee(t_and_or *andor, int8_t lr)
+uint8_t		condition_respected(t_and_or *andor, int8_t lr)
 {
 	if (andor->type == AND_IF)
 		return (lr == SUCCESS ? 1 : 0);
@@ -15,37 +13,36 @@ int		condition_respectee(t_and_or *andor, int8_t lr)
 	return (FAILURE);
 }
 
-int		lvl_simple_cmd(t_cfg *shell, t_list *s_cmd, char *cmd, uint8_t fg)
+uint8_t		lvl_simple_cmd(t_cfg *shell, t_list *s_cmd, char *cmd, uint8_t fg)
 {
-	uint8_t		tmp;
+	uint8_t		ret_job;
 	t_job		job;
 
 	cmd_to_job(shell, &job, s_cmd, cmd);
 	if ((job.fg = fg))
 		set_termios(TCSADRAIN, &job.term_eval);
-	run_job(shell, &job, job.process);
-	tmp = job.ret;				//en attendant var intern? 
-	routine_clean_job(&job);
-	return (tmp);
+	ret_job = run_job(shell, &job, job.process);
+	routine_clean_job(&job, sizeof(t_job));
+	return (ret_job);
 }
 
-int		lvl_and_or(t_cfg *shell, t_list *lst)
+uint8_t		lvl_and_or(t_cfg *shell, t_list *lst)
 {
-	uint8_t		lr;				//en attendant var intern ?
+	uint8_t		ret_job;
 	t_and_or	*andor;
 
 	if (!lst)
 		return (0);
 	andor = lst->data;
-	lr = lvl_simple_cmd(shell, andor->s_cmd, andor->str, !andor->background);
-	if (lst->next && condition_respectee(andor, lr))
+	ret_job = lvl_simple_cmd(shell, andor->s_cmd, andor->str, !andor->background);
+	if (lst->next && condition_respected(andor, ret_job))
 		lvl_and_or(shell, lst->next); //continue
 	else if (lst->next)
 		lvl_and_or(shell, lst->next->next);
 	return (0);
 }
 
-int		lvl_cmd_table(t_cfg *shell, t_list *lst)
+void		lvl_cmd_table(t_cfg *shell, t_list *lst)
 {
 	t_cmd_table		*cmd;
 
@@ -55,19 +52,18 @@ int		lvl_cmd_table(t_cfg *shell, t_list *lst)
 		lvl_and_or(shell, cmd->and_or);
 		lst = lst->next;
 	}
-	return (0);
 }
 
-int		ft_eval(t_list *cmd_table)
+uint8_t		ft_eval(t_list *cmd_table)
 {
 	t_cfg *shell;
 
 	shell = cfg_shell();
 	if (shell->debug)
-		printf("\n\n----------- eval -----------\n\n\n\n");
+		ft_printf("\n\n----------- eval -----------\n\n\n\n");
 	set_signal_ign();
 	lvl_cmd_table(shell, cmd_table);
 	if (shell->debug)
-		printf("----------- eval -----------\n\n");
+		ft_printf("----------- eval -----------\n\n");
 	return (0);
 }
