+++ srcs/evaluator/tools_termios.c
@@ -3,7 +3,7 @@
 #include <termios.h>
 #include <unistd.h>
 
-void		set_termios(int32_t mode, struct termios *term)
+void	set_termios(int32_t mode, struct termios *term)
 {
 	if (tcsetattr(STDIN_FILENO, mode, term) == -1)
 		perror("[RUN JOB] error tcsetattr");
